Heat map technical skills correlation
Here is a heat map showing the correlation for the top 10 mentioned skills. The heat map itself is not ordered by number of mentions, but rather by how much it correlates to the rest of the columns. For instance, SQL is in the first position, meaning that the people that know SQL tend to know more programs than the people that know Matlab.

It is still far from being as pretty as it can be, but I played a bit with parameters to try to make it as neat as possible.